requirements.txt
requests
beautifulsoup4
feedparser
google-generativeai
pyyaml
